# scripts/openapi_schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cadence v2",
    "description": "\n# Multi-Tenant AI Orchestration Platform\n\nProduction-ready framework for deploying AI agent orchestrators at scale.\n\n## Features\n\n* **Multi-Backend Support**: LangGraph, OpenAI Agents SDK, Google ADK\n* **Multi-Tenancy**: Complete tenant isolation with BYOK (Bring Your Own Key)\n* **Three Orchestration Modes**: Supervisor, Coordinator, Handoff\n* **Hot-Reload**: Configuration changes without restart\n* **Scalable Architecture**: Hot/Warm/Cold tier pool (1000+ instances)\n* **Streaming**: Server-Sent Events (SSE) for real-time responses\n\n## Authentication\n\n### JWT Authentication\n```\nAuthorization: Bearer <jwt_token>\n```\n\n### API Key Authentication\n```\nX-API-Key: <api_key>\n```\n\n## Rate Limiting\n\nAll API endpoints are rate-limited per tenant. Default: 100 requests/minute.\n\n## Documentation\n\n* Platform Docs: [README](https://github.com/jonaskahn/cadence)\n* SDK Docs: [cadence-sdk](https://github.com/jonaskahn/cadence-sdk)\n",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    },
    {
      "url": "https://api.example.com",
      "description": "Production server"
    },
    {
      "url": "https://staging-api.example.com",
      "description": "Staging server"
    }
  ],
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "description": "Authenticate and receive a JWT.\n\nThe JWT carries only sub (user_id) and jti (ULID session key).\nOrg membership and role data live in Redis and are resolved per request.\n\nArgs:\n    login_request: Username and password\n    request: FastAPI request\n\nReturns:\n    LoginResponse with signed JWT\n\nRaises:\n    HTTPException: 401 if credentials are invalid",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/me/profile": {
      "patch": {
        "tags": [
          "auth"
        ],
        "summary": "Update My Profile",
        "description": "Update the authenticated user's own password.\n\nUsername and email cannot be changed via this endpoint.\n\nArgs:\n    update_request: Current and new password\n    request: FastAPI request\n    context: Authenticated context\n\nRaises:\n    HTTPException: 401 if current password is wrong",
        "operationId": "update_my_profile_api_me_profile_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/plugins": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List System Plugins",
        "description": "List all system plugins, optionally filtered by tag.\n\nArgs:\n    request: FastAPI request\n    tag: Optional tag filter\n    tenant_context: Tenant context from JWT\n\nReturns:\n    List of system plugin metadata",
        "operationId": "list_system_plugins_api_admin_plugins_get",
        "parameters": [
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SystemPluginResponse"
                  },
                  "title": "Response List System Plugins Api Admin Plugins Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/settings/{key}": {
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "Update Global Setting",
        "description": "Update global setting (Tier 2).\n\nArgs:\n    key: Setting key\n    update_request: New value\n    request: FastAPI request\n    tenant_context: Admin tenant context\n\nReturns:\n    Updated setting\n\nRaises:\n    HTTPException: If operation fails",
        "operationId": "update_global_setting_api_admin_settings__key__patch",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGlobalSettingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalSettingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GlobalSettingResponse": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "value": {
            "title": "Value"
          },
          "value_type": {
            "type": "string",
            "title": "Value Type"
          },
          "description": {
            "type": "string",
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "key",
          "value",
          "value_type",
          "description"
        ],
        "title": "GlobalSettingResponse",
        "description": "Global setting response (Tier 2).\n\nAttributes:\n    key: Setting key\n    value: Setting value\n    value_type: Value type\n    description: Setting description"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "minLength": 1,
            "title": "Username",
            "description": "Username"
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "title": "Password",
            "description": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest",
        "description": "Login credentials.\n\nAttributes:\n    username: Account username\n    password: Plain-text password"
      },
      "LoginResponse": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "token"
        ],
        "title": "LoginResponse",
        "description": "Successful login response.\n\nAttributes:\n    token: Signed JWT; use as 'Authorization: Bearer <token>'"
      },
...
      "SystemPluginResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "pid": {
            "type": "string",
            "title": "Pid"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "tag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag"
          },
          "is_latest": {
            "type": "boolean",
            "title": "Is Latest"
          },
          "s3_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "S3 Path"
          },
          "default_settings": {
            "additionalProperties": true,
            "type": "object",
            "title": "Default Settings",
            "default": {}
          },
          "capabilities": {
            "items": {},
            "type": "array",
            "title": "Capabilities",
            "default": []
          },
          "agent_type": {
            "type": "string",
            "title": "Agent Type"
          },
          "stateless": {
            "type": "boolean",
            "title": "Stateless"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "pid",
          "version",
          "name",
          "is_latest",
          "agent_type",
          "stateless",
          "is_active"
        ],
        "title": "SystemPluginResponse",
        "description": "System plugin response model."
      },
...
      "UpdateGlobalSettingRequest": {
        "properties": {
          "value": {
            "title": "Value",
            "description": "New setting value"
          }
        },
        "type": "object",
        "required": [
          "value"
        ],
        "title": "UpdateGlobalSettingRequest",
        "description": "Update global setting request.\n\nAttributes:\n    value: New value"
      },
...
      "UpdateProfileRequest": {
        "properties": {
          "current_password": {
            "type": "string",
            "minLength": 1,
            "title": "Current Password",
            "description": "Current password"
          },
          "new_password": {
            "type": "string",
            "minLength": 8,
            "title": "New Password",
            "description": "New password (min 8 chars)"
          }
        },
        "type": "object",
        "required": [
          "current_password",
          "new_password"
        ],
        "title": "UpdateProfileRequest",
        "description": "Self-service profile update.\n\nOnly password can be changed. Username and email are immutable via this endpoint.\n\nAttributes:\n    current_password: Current password for verification\n    new_password: New password to set"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from authentication endpoint. Format: `Bearer <token>`"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for service-to-service authentication. Format: `X-API-Key: <key>`"
      }
    }
  },
...
  "security": [
    {
      "BearerAuth": []
    },
    {
      "ApiKeyAuth": []
    }
  ],
...
}